Add undo/redo functions to Widgets
Some of the functions may need to be changed depending on how
we opt to call the function responsible for opening the image
gallery. If we opted to have this function be called from outside
the `ImageWidget`, the current code is fine. However, if we
opt for the opposite than we need to repurpose the current
function responsible for updating a `ImageWidget` image.