Implement PaletteViewController main features and prepares it to integration ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
PaletteViewController may suffer some additional changes based on refactoring of code and adaptations to integrate with the project. Bear in mind that the protocols and structs created in its file will probably need to be refactored and should not be taken as a ready for use. The same can be said from it's item drag implementation, as it still needs further definition to comply with interactions in and out of the app.
2 files were added:
- CarouselCollectionView.swift
--A subclass of UICollectionView, it implements a custom flow layout that create a single line drag enabled collectionView, with cells filling up a big space on the screen.
It registers a special subclass of UICollectionViewCell, called CarouselViewCell, with the identifier "default".

- CarouselCellView.swift
-- A subclass of UICollectionViewCell, it adds an UIImageView to the standard cell, as well as some settings for the same.