Integrate save/load with Container
Integrated save/load using JSON with the `ContainerViewController`
class. Unfortunately, due to some issues with our architecture
of choice, we had to increase a lot of the coupling.

Another severe issue that should be resolved in a later update/rework
of the app, is the way in which we get the current widgets' frames.
As a result of UI bein only accessible in the main thread, we had to
create another fake, onyl for saving, frame that we called
`internalFrame`.